Use snake_case for identifiers
This is an opinionated commit. The Ruby community's normative code style uses lowercase snake_case for most identifiers. I like this style since it makes it easier to read than having words compacted together.

For objects that hold nothing but functions I use CamelCase (but there were none in the code base to use as example).