Avoid scrollbars on Chrome
Chrome seems to ignore hiding scrollbars and the #app div was set to the width needed to show the map part in order to make the layout correct, whether the map were shown or not. With these changes a mapless class is added to #app which limits the width to 448px.