Handle data as immutable objects
This commit introduce the immutability-helper library which makes it easier to change data by creating copies rather than mutating data. This in turn will help with refactoring into the approach mandated by ReactJs.

There is one necessary change to the library in that copies of objects retain their __proto__ rather than their constructor. This way it doesn't matter if the object were created with a proper prototype (as in our case), or through a constructor function.