Implement dungeon maps with location availability
Dungeon maps are shown by clicking their square on the map, and dismissed again by pressing escape. Clicking dungeon locations or doors on the map marks them, just as overworld locations. Marking doors count toward key logic, and by marking locations the chest count and boss state is updated on the item tracker.

Note that decreasing the chest count on the item tracker such that the count mismatch with the count of marked dungeon locations shows yellow state for the whole dungeon, leaving it to the player to figure out how to proceed. The boss and its boss location is except from this since they correspond to each other uniquely.

The launcher have two new options for keysanity to indicate that the player know ice palace bomb jumps, and palace of darkness hammeryump respectively. Without the ice bomb jump option "east wing" and the boss locations might show as yellow to indicate possibly requiring bomb jumps.

- SW is not guaranteed to be progressable unless, besides firerod, one also have more than one unique item left, or a sword.